Fixed nsedit webUI issue
After pulling down the latest version of nsedit the webUI was broken and not allowing
anyone to log in. A change to the template for config.inc.php was required as well as
changing the path for config.inc.php to be within includes/ which was the way it was
to be for some time. However it was still working in the root of nsedit. But this has been
resolved.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>